refactor: 파일 업로드 로직 수정 (#128)

+++ src/main/java/com/kcc/pms/domain/output/domain/vo/FileMasterVO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class FileMasterVO {
-    private Integer fl_ms_no;
+    private Long fl_ms_no;
     private String fl_cd;
     private String use_yn;
